move zipcode function to user.js, if adding address, get new state and district
